finished updating routes

+++ routes/expensesRoutes.js
@@ -2,23 +2,15 @@ const express = require("express");
 const router = express.Router();
 const {Expense} = require("../models");
 
-// GET ALL
 router.get("/", async (req, res) => {
-try {
-    const data = await Expense.findAll();
-    
-    
-    expenses = data.map(function(expenseObj){
-        return expenseObj.toJSON()
-    })
-    console.log(expenses)
-    // res.json(data);
-    res.render('landing', {expenses:expenses})
-} catch (err) {
-    console.log(err);
-    res.status(500).json({ msg: "error occurred", err });
-}
-});
+    try {
+        const data = await Expense.findAll();
+        res.json(data);
+    } catch (err) {
+        console.log(err);
+        res.status(500).json({ msg: "error occurred", err });
+    }
+    });
 
 // GET BY ID
 router.get("/:id", (req, res) => {
@@ -54,39 +46,54 @@ router.post('/', async (req, res) => {
     }
 });
 
+//TODO: does not block user from editing other user's expenses
+// PUT (EDIT BY ID)
+
+router.put('/:id', async (req, res) => {
+    try {
+        const updateExpense = await Expense.findByPk(req.params.id);
+        if(updateExpense.user_id!==req.session.user.id){
+            return res.status(403).json({msg:"not your expense!"})
+        }
+        const data = await Expense.update(req.body, {
+            where: {
+                expense_id: req.params.id,
+            },
+        });
+        if (data[0] === 0) {
+            return res.status(404).json({ msg: "does not exist!" });
+        }
+        res.json(data);
+    }
+    catch (err) {
+        console.log(err);
+        res.status(500).json({ msg: "error occurred", err });
+    }
+}
+);
 
 // DELETE
-router.delete("/:id", (req, res) => {
-Expense.destroy({
-    where: {
-    expense_id: req.params.id,
-    },
-}).then((data) => {
-    if(data===0){
-    return res.status(404).json({msg:"does not exist!"})
+router.delete("/:id", async (req, res) => {
+    try {
+        const deleteExpense = await Expense.findByPk(req.params.id);
+        if(deleteExpense.user_id!==req.session.user.id){
+            return res.status(403).json({msg:"not your expense!"})
+        }
+        const data = await Expense.destroy({
+            where: {
+                expense_id: req.params.id,
+            },
+        });
+        if (data === 0) {
+            return res.status(404).json({ msg: "does not exist!" });
+        }
+        res.json(data);
     }
-    res.json(data);
-}).catch(err=>{
-    console.log(err);
-    res.status(500).json({msg:"error occurred",err})
-});
-});
-
-// PUT (EDIT BY ID)
-router.put("/:id", (req, res) => {
-Expense.update(req.body, {
-    where: {
-    expense_id: req.params.id,
-    },
-}).then((data) => {
-    if(data[0]===0){
-    return res.status(404).json({msg:"does not exist!"})
+    catch (err) {
+        console.log(err);
+        res.status(500).json({ msg: "error occurred", err });
     }
-    res.json(data);
-}).catch(err=>{
-    console.log(err);
-    res.status(500).json({msg:"error occurred",err})
-});
-});
+}
+);
 
 module.exports = router;
